requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
supabase==2.1.0
yake==0.4.8
Pillow==10.1.0
numpy==1.24.3
requests==2.31.0
python-docx==1.1.0
PyPDF2==3.0.1
python-multipart==0.0.6
setuptools==69.5.1
torch
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2